+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   Minor: Improved logging. ([#8](https://github.com/twesterapp/twester/pull/8))
 -   Minor: Removed menu bar. ([#11](https://github.com/twesterapp/twester/pull/11))
+-   Minor: Show `IconButton` with `IconEye` for password variant of `InputText`. ([#12](https://github.com/twesterapp/twester/pull/12))
 
 ## 1.0.2 (08-10-2021)